+++ main.lua
@@ -4,7 +4,7 @@ local snax = require "skynet.snax"
 local is_windows = package.config:sub(1,1) == '\\'
 
 skynet.start(function()
-	skynet.error("FreeIOE@Skynet Starting")
+	skynet.error("FreeIOE Starting...")
 	--skynet.newservice("exec_sal")
 	if not is_windows and not os.getenv("IOE_RUN_AS_DAEMON") then
 		local console = skynet.newservice("console")