+++ snax/logger.lua
@@ -1,4 +1,3 @@
-local skynet = require 'skynet'
 local snax = require 'skynet.snax'
 local ioe = require 'ioe'